(Schema) refactored user and task schema
