+++ Schema/__init__.py
@@ -1,8 +1,7 @@
-from Schema.UserLoginSchema import UserLoginSchema
-from Schema.UserCreateSchema import UserCreateSchema
-from Schema.TaskCreateSchema import TaskCreateSchema
+from Schema.user import UserLoginSchema, UserCreateSchema
+from Schema.task import TaskSchema
 from Schema.auth import GoogleUserDataSchema, YandexUserDataSchema
 
 __all__ = ["UserLoginSchema", "UserCreateSchema",
-           "TaskCreateSchema", "GoogleUserDataSchema",
+           "TaskSchema", "GoogleUserDataSchema",
            "YandexUserDataSchema"]
