+++ Schema/task.py
@@ -1,6 +1,6 @@
 from pydantic import BaseModel
 
-class TaskCreateSchema(BaseModel):
+class TaskSchema(BaseModel):
     id : int
     title : str
     user_id : int
