+++ Schema/user.py
@@ -10,3 +10,14 @@ class UserCreateSchema(BaseModel):
 
     class Config:
         from_attributes = True
+
+
+class UserLoginSchema(BaseModel):
+    user_id : int
+    access_token: str
+
+    class Config:
+        from_attributes = True
+
+    def __str__(self):
+        return f"id - {self.user_id}  токен - {self.access_token}"
