+++ service/TaskService.py
@@ -1,7 +1,7 @@
 from dataclasses import dataclass
 
-from Schema.TaskCreateSchema import TaskCreateSchema
-from models import Task
+from Schema.task import TaskSchema
+
 from repository import TaskRepository
 
 
@@ -10,11 +10,11 @@ class TaskService:
 
     task_repository: TaskRepository
 
-    def get_tasks(self, user_id : int) -> list[TaskCreateSchema]:
+    def get_tasks(self, user_id : int) -> list[TaskSchema]:
         tasks = self.task_repository.get_all_user_tasks(user_id=user_id)
-        task_list = [TaskCreateSchema.model_validate(task) for task in tasks]
+        task_list = [TaskSchema.model_validate(task) for task in tasks]
         return task_list
 
-    def create_task(self, title : str, user_id : int) -> TaskCreateSchema:
+    def create_task(self, title : str, user_id : int) -> TaskSchema:
         task = self.task_repository.create_task(title=title, user_id=user_id)
-        return TaskCreateSchema(id=task.id, title=task.title, user_id=user_id)
+        return TaskSchema(id=task.id, title=task.title, user_id=user_id)
